Restructure for pushdowns
