+++ spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/SpannerTableSchema.java
@@ -17,6 +17,7 @@ import com.google.cloud.spanner.ResultSet;
 import com.google.cloud.spanner.Statement;
 import com.google.cloud.spanner.Struct;
 import com.google.cloud.spanner.connection.Connection;
+import com.google.cloud.spark.spanner.scan.SpannerTable;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Locale;
