+++ spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/graph/SpannerGraphScanner.java
@@ -28,12 +28,12 @@ import com.google.cloud.spanner.Statement;
 import com.google.cloud.spanner.TimestampBound;
 import com.google.cloud.spark.spanner.BatchClientWithCloser;
 import com.google.cloud.spark.spanner.SpannerInputPartitionContext;
-import com.google.cloud.spark.spanner.SpannerPartition;
-import com.google.cloud.spark.spanner.SpannerPartitionReaderFactory;
 import com.google.cloud.spark.spanner.SpannerRowConverter;
-import com.google.cloud.spark.spanner.SpannerScanner;
 import com.google.cloud.spark.spanner.SpannerUtils;
 import com.google.cloud.spark.spanner.graph.query.SpannerGraphQuery;
+import com.google.cloud.spark.spanner.scan.SpannerPartition;
+import com.google.cloud.spark.spanner.scan.SpannerPartitionReaderFactory;
+import com.google.cloud.spark.spanner.scan.SpannerScanner;
 import com.google.common.collect.ImmutableSet;
 import io.grpc.Context;
 import io.grpc.MethodDescriptor;
