+++ spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/graph/query/ElementTableQuery.java
@@ -3,13 +3,13 @@ package com.google.cloud.spark.spanner.graph.query;
 import com.google.cloud.Tuple;
 import com.google.cloud.spanner.Statement;
 import com.google.cloud.spark.spanner.SpannerRowConverter;
-import com.google.cloud.spark.spanner.SpannerTable;
 import com.google.cloud.spark.spanner.SpannerTableSchema;
 import com.google.cloud.spark.spanner.graph.PropertyGraph;
 import com.google.cloud.spark.spanner.graph.PropertyGraph.GraphElementTable;
 import com.google.cloud.spark.spanner.graph.PropertyGraph.GraphPropertyDefinition;
 import com.google.cloud.spark.spanner.graph.SpannerGraphConfigs.LabelConfig;
 import com.google.cloud.spark.spanner.graph.SpannerRowConverterWithSchema;
+import com.google.cloud.spark.spanner.scan.SpannerTable;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.Arrays;
