+++ spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/scan/SpannerPartition.java
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.cloud.spark.spanner;
+package com.google.cloud.spark.spanner.scan;
 
+import com.google.cloud.spark.spanner.InputPartitionContext;
 import java.io.Serializable;
 import org.apache.spark.Partition;
 import org.apache.spark.sql.catalyst.InternalRow;
