+++ spark-3.1-spanner-lib/src/test/java/com/google/cloud/spark/spanner/integration/SpannerInputPartitionReaderContextIntegrationTestBase.java
@@ -27,12 +27,12 @@ import com.google.cloud.spanner.TimestampBound;
 import com.google.cloud.spark.spanner.InputPartitionReaderContext;
 import com.google.cloud.spark.spanner.SpannerInputPartitionContext;
 import com.google.cloud.spark.spanner.SpannerInputPartitionReaderContext;
-import com.google.cloud.spark.spanner.SpannerPartitionReader;
 import com.google.cloud.spark.spanner.SpannerRowConverterDirect;
-import com.google.cloud.spark.spanner.SpannerScanBuilder;
-import com.google.cloud.spark.spanner.SpannerScanner;
-import com.google.cloud.spark.spanner.SpannerTable;
 import com.google.cloud.spark.spanner.SpannerUtils;
+import com.google.cloud.spark.spanner.scan.SpannerPartitionReader;
+import com.google.cloud.spark.spanner.scan.SpannerScanBuilder;
+import com.google.cloud.spark.spanner.scan.SpannerScanner;
+import com.google.cloud.spark.spanner.scan.SpannerTable;
 import java.io.IOException;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
