+++ spark-3.1-spanner-lib/src/test/java/com/google/cloud/spark/spanner/integration/SpannerTestBase.java
@@ -29,10 +29,10 @@ import com.google.cloud.spanner.Spanner;
 import com.google.cloud.spanner.SpannerException;
 import com.google.cloud.spanner.SpannerOptions;
 import com.google.cloud.spanner.Statement;
-import com.google.cloud.spark.spanner.SpannerTable;
 import com.google.cloud.spark.spanner.SpannerTestUtils;
 import com.google.cloud.spark.spanner.SpannerUtils;
 import com.google.cloud.spark.spanner.TestData;
+import com.google.cloud.spark.spanner.scan.SpannerTable;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import io.micrometer.observation.Observation.CheckedRunnable;
