+++ spark-3.1-spanner-lib/src/test/java/com/google/cloud/spark/spanner/scan/SpannerScannerTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.cloud.spark.spanner;
+package com.google.cloud.spark.spanner.scan;
 
 import static com.google.common.truth.Truth.assertThat;
